Copy code that changes the PowerShell execution policy to test projects that need it
